refactor: create new controller to  sepreate cocktail and CocktailSong controller logic
